Moved analytics from services to processors
