+++ packages/gafl-webapp-service/src/processors/__tests__/analytics.spec.js
