+++ packages/gafl-webapp-service/src/processors/analytics.js
